+++ src/server/game/Spells/SpellScript.h
@@ -486,8 +486,8 @@ class AuraScript : public _SpellScript
         typedef void(CLASSNAME::*AuraEffectPeriodicFnType)(AuraEffect const*); \
         typedef void(CLASSNAME::*AuraEffectUpdateFnType)(uint32, AuraEffect*); \
         typedef void(CLASSNAME::*AuraEffectUpdatePeriodicFnType)(AuraEffect*); \
-        typedef void(CLASSNAME::*AuraEffectCalcAmountFnType)(AuraEffect const*, int32 &, bool &); \
-        typedef void(CLASSNAME::*AuraEffectChangeTickDamageFnType)(AuraEffect const*, int32 &, Unit*, bool); \
+        typedef void(CLASSNAME::*AuraEffectCalcAmountFnType)(AuraEffect const*, float &, bool &); \
+        typedef void(CLASSNAME::*AuraEffectChangeTickDamageFnType)(AuraEffect const*, float &, Unit*, bool); \
         typedef void(CLASSNAME::*AuraEffectCalcPeriodicFnType)(AuraEffect const*, bool &, int32 &); \
         typedef void(CLASSNAME::*AuraEffectCalcSpellModFnType)(AuraEffect const*, SpellModifier* &); \
         typedef void(CLASSNAME::*AuraEffectAbsorbFnType)(AuraEffect*, DamageInfo &, uint32 &); \
@@ -558,7 +558,7 @@ class AuraScript : public _SpellScript
         {
             public:
                 EffectCalcAmountHandler(AuraEffectCalcAmountFnType _pEffectHandlerScript, uint8 _effIndex, uint16 _effName);
-                void Call(AuraScript* auraScript, AuraEffect const* aurEff, int32 & amount, bool & canBeRecalculated);
+                void Call(AuraScript* auraScript, AuraEffect const* aurEff, float & amount, bool & canBeRecalculated);
             private:
                 AuraEffectCalcAmountFnType pEffectHandlerScript;
         };
@@ -574,7 +574,7 @@ class AuraScript : public _SpellScript
         {
             public:
                 EffectChangeTickDamageHandler(AuraEffectChangeTickDamageFnType _pEffectHandlerScript, uint8 _effIndex, uint16 _effName);
-                void Call(AuraScript* auraScript, AuraEffect const* aurEff, int32 & amount, Unit* target, bool crit);
+                void Call(AuraScript* auraScript, AuraEffect const* aurEff, float & amount, Unit* target, bool crit);
             private:
                 AuraEffectChangeTickDamageFnType pEffectHandlerScript;
         };
