+++ src/server/scripts/Kalimdor/CavernsOfTime/BattleForMountHyjal/hyjal_trash.cpp
@@ -1456,7 +1456,7 @@ public:
                     EnterEvadeMode();
                     return;
                 }
-                int dmg = 500+rand()%700;
+                float dmg = 500 + rand() % 700;
                 me->CastCustomSpell(me->getVictim(), SPELL_EXPLODING_SHOT, &dmg, 0, 0, false);
                 ExplodeTimer = 5000+rand()%5000;
             } else ExplodeTimer -= diff;
