+++ src/server/scripts/Outland/BlackTemple/boss_reliquary_of_souls.cpp
@@ -551,7 +551,7 @@ public:
             }
             else
             {
-                int32 bp0 = damage / 2;
+                float bp0 = damage / 2;
                 me->CastCustomSpell(done_by, AURA_OF_DESIRE_DAMAGE, &bp0, NULL, NULL, true);
             }
         }
